Fix servicios

+++ src/main/java/com/nuvu/customer/repository/CustomerRepository.java
@@ -1,13 +1,11 @@
 package com.nuvu.customer.repository;
 
-import java.math.BigDecimal;
-
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import com.nuvu.customer.entity.CustomerEntity;
 
 @Repository
-public interface CustomerRepository extends JpaRepository<CustomerEntity, BigDecimal> {
+public interface CustomerRepository extends JpaRepository<CustomerEntity, Integer> {
 
 }
